fix psql schema syntax

+++ tpch-2.14.0/experiments/final-schema/psql/customer-enc.sql
@@ -1,8 +1,8 @@
 CREATE TABLE customer_enc ( 
-    c_custkey_DET     integer  not null,
+    c_custkey_DET     bigint  not null,
     c_name_DET        bytea not null,
     c_address_DET     bytea not null,
-    c_nationkey_DET   integer  not null,
+    c_nationkey_DET   bigint  not null,
     c_phone_DET       bytea not null,
     c_acctbal_DET     bigint  not null,
     c_acctbal_OPE     bytea not null,
